+++ wp-content/themes/falcon-tailwind-theme-starter/assets/js/editor-styles.js
@@ -2,34 +2,10 @@
 
 wp.domReady( () => {
 
-
-	// Alternative Button Styles
+	// This is example of how to add a block style from js
 	wp.blocks.registerBlockStyle( 'core/button', {
-		name: 'alternative',
-		label: 'Alternative'
+		name: 'ExampleButton',
+		label: 'ExampleButton'
 	} );
-
-	wp.blocks.registerBlockStyle( 'core/button', {
-		name: 'alternative-outline',
-		label: 'Alternative Outline'
-	} );
-
-	// CTA Button Styles
-	wp.blocks.registerBlockStyle( 'core/button', {
-		name: 'cta',
-		label: 'CTA'
-	} );
-
-	wp.blocks.registerBlockStyle( 'core/button', {
-		name: 'cta-outline',
-		label: 'CTA Outline'
-	} );
-
-	// Button as link
-	wp.blocks.registerBlockStyle( 'core/button', {
-		name: 'link',
-		label: 'Link'
-	} );
-	
 
 } );