+++ wp-content/themes/falcon-tailwind-theme-starter/assets/js/editor-variations.js
@@ -3,30 +3,23 @@
 wp.domReady( () => {
 
 	wp.blocks.registerBlockVariation( 'core/spacer', {
-		name: 'small',
-		title: 'Spacer: Small',
+		name: 'section/spacer',
+		title: 'Section spacer',
 		attributes: {
-			height: '2rem'
+			// add css class
+			className: 'section-spacer',
+			height: 'var:preset|spacing|8',
 		},
-		scope: [ 'block', 'inserter', 'transform' ]
-	} );
+			scope: [ 'block', 'inserter', 'transform' ]
+		} 
+	);	
 
 	wp.blocks.registerBlockVariation( 'core/spacer', {
 		name: 'medium',
 		title: 'Spacer: Medium',
 		attributes: {
-			height: 'clamp(2rem, calc(2rem + ((1vw - 0.4rem) * 8.3333)), 4rem)'
+			height: 'clamp(1.875rem, calc(1.875rem + ((1vw - 0.4rem) * 8.3333)), 3rem)'
 		},
 		scope: [ 'block', 'inserter', 'transform' ]
 	} );
-
-	wp.blocks.registerBlockVariation( 'core/spacer', {
-		name: 'large',
-		title: 'Spacer: Large',
-		attributes: {
-			height: 'clamp(4rem, calc(4rem + ((1vw - 0.4rem) * 16.6667)), 8rem)'
-		},
-		scope: [ 'block', 'inserter', 'transform' ]
-	} );
-
 } );