+++ wp-content/themes/falcon-tailwind-theme-starter/browsersync-config.js
@@ -1,14 +1,25 @@
 module.exports = {
- 
-    "files": ["*.css", "*.js", "*.php", "./**/*.php",, "*.html", "./**/*.html", "./**/*.js", "./**/*.css"],
-    "reloadOnRestart": true,
-    "port": 3000,
-    "browser": "chrome",
-    "reloadDelay": 100,
-    "watchEvents": [ 
-        "change" 
+    files: [
+      "theme.json",
+      "assets/**/*.*",
+      "blocks/**/*.{php,js,css,json}",
+      "includes/**/*.php",
+      "functions.php",
+      "patterns/**/*.{html,php}",
+      "parts/**/*.html",
+      "templates/**/*.{html,php}",
+      "styles/**/*.json",
+      "safelist.txt",
+      "tailwind.config.js",
     ],
-    "proxy": {
-        target: "https://falcon-tailwind-wordpress-starter.local/" // Enter your local WP URL here
+    reloadOnRestart: true,
+    port: 3000,
+    browser: "chrome",
+    reloadDelay: 200, // Wait 100 milliseconds before triggering a reload event
+    reloadDebounce: 2000, // Wait 1 second after the last reload event has been triggered before reloading the page
+    watchEvents: ["change"],
+    proxy: {
+      target: "https://falcon-tailwind-wordpress-starter.local/", // Enter your local WP URL here
     },
-};
+  };
+