+++ wp-content/themes/falcon-tailwind-theme-starter/resources/js/app.js
@@ -1,11 +1,1 @@
-import 'flowbite';
-
-
-// Navigation toggle
-window.addEventListener('load', function () {
-      let main_navigation = document.querySelector('#primary-menu');
-      document.querySelector('#primary-menu-toggle').addEventListener('click', function (e) {
-            e.preventDefault();
-            main_navigation.classList.toggle('hidden');
-      });
-});
+// HERE YOU CAN ADD YOUR CUSTOM JS CODE